+++ birdnet-go/rootfs/etc/cont-init.d/99-run.sh
@@ -9,14 +9,22 @@ set -e
 bashio::log.info "ALSA_CARD option is set to $(bashio::config "ALSA_CARD"). If the microphone doesn't work, please adapt it"
 echo " "
 
+# Adjust microphone volume if needed
+if command -v amixer >/dev/null 2>/dev/null; then
+  current_volume="$(amixer sget Capture | grep -oP '\[\d+%\]' | tr -d '[]%' | head -1 2>/dev/null || echo "100")" || true
+  if [[ "$current_volume" -eq 0 ]]; then
+      amixer sset Capture 70%
+      bashio::log.warning "Microphone was off, volume set to 70%."
+  fi || true
+fi
+
 ########################
 # CONFIGURE birdnet-go #
 ########################
 
 bashio::log.info "Starting app..."
-COMMAND="$(bashio::config "COMMAND")"
 # shellcheck disable=SC2086
-/usr/bin/birdnet-go $COMMAND & true
+/usr/bin/entrypoint.sh birdnet-go realtime & true
 
 # Wait for app to become available to start nginx
 bashio::net.wait_for 8080 localhost 900