+++ birdnet-go/CHANGELOG.md
@@ -1,5 +1,10 @@
+## 20250427-7 (15-05-2025)
+- Breaking change: COMMAND addon option removed. Please instead use the config.yaml to define the RTSP feeds
+- Use entrypoint
+
 ## 20250427-2 (27-04-2025)
 - Minor bugs fixed
+
 ## 20250427 (27-04-2025)
 - Minor bugs fixed