+++ nextcloud/CHANGELOG.md
@@ -1,6 +1,11 @@
+## 31.0.4-3 (15-05-2025)
+- Minor bugs fixed
+## 31.0.4-2 (15-05-2025)
+- Disable upload file limits
 
 ## 31.0.4 (19-04-2025)
 - Update to latest version from linuxserver/docker-nextcloud (changelog : https://github.com/linuxserver/docker-nextcloud/releases)
+
 ## 31.0.2-2 (23-03-2025)
 - Fix permissions check